requirements.txt
aerich
aiogram
asyncpg
betterlogging
pydantic-settings
redis
tortoise-orm
pydantic
pydantic_core
Booktype